+++ src/plugin.hpp
@@ -19,6 +19,7 @@ extern Model* modelBolt;
 extern Model* modelInfix;
 extern Model* modelInfixMicro;
 extern Model* modelStrip;
+extern Model* modelStripCon4;
 extern Model* modelEightFace;
 extern Model* modelEightFaceX2;
 extern Model* modelMidiCat;